requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
tqdm>=4.66.0
git+https://github.com/jcarlosroldan/procgen
